src/requirements.txt
Django==5.0.9
requests==2.32.3
requests-mock==1.10.0
Authlib==1.3.1
python-dotenv==1.0.1
djangorestframework==3.15.2
markdown==3.7      # Markdown support for the browsable API.
django-filter==24.3
jsonpatch==1.33
mysqlclient==2.2.7
factory_boy
